requirements.txt
Django==4.2.16
django-cors-headers==4.6.0
django-graphql-jwt==0.4.0
graphene-django==3.2.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
django-filter==24.3